+++ packages/cli-kube/src/handlers/index.js
@@ -3,11 +3,13 @@
 //
 
 export * from './assemble';
+export * from './cert';
 export * from './delete';
 export * from './deploy';
 export * from './get';
 export * from './install';
 export * from './list';
+export * from './local';
 export * from './register';
 export * from './otp';
 export * from './start';